db-parser: extract PDBLookupParams
